requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
authlib==1.2.0
aiohttp==3.8.5
ecdsa==0.18.0
asn1crypto==1.5.1
python-dotenv==1.0.0
starlette==0.27.0
